+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Command parameters can now be grouped with double quotes. eg. `<<add-recipe "Banana Sensations">` and `<<move "My Game Object" LocationName>>` (@andiCR)
 
+- You can now add dialogue views to dialogue runner at any time.
+
+- The inspector for Yarn scripts now allows you to change the Project that the script belongs to. (@radiatoryang)
+
 ### Changed
 
 - Updated to support new error handling in Yarn Spinner. 
@@ -26,6 +30,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 * `InMemoryVariableStorage` now throws an exception if you attempt to get or set a variable whose name doesn't start with `$`.
+* `LineView` now has an onCharacterTyped event that triggers for each character typed in the typewriter effect.
 
 ### Changed